refactor: clean up the number of hook files
