+++ packages/react-dnd/src/hooks/internal/drag.ts
@@ -8,7 +8,18 @@ import {
 import { DragDropMonitor, DragSource } from 'dnd-core'
 import registerSource from '../../registerSource'
 import { useDragDropManager } from './useDragDropManager'
+import DragSourceMonitorImpl from '../../DragSourceMonitorImpl'
+import createSourceConnector from '../../createSourceConnector'
 const invariant = require('invariant')
+
+export function useDragSourceMonitor(): [DragSourceMonitor, any] {
+	const manager = useDragDropManager()
+	const monitor = useMemo(() => new DragSourceMonitorImpl(manager), [manager])
+	const connector = useMemo(() => createSourceConnector(manager.getBackend()), [
+		manager,
+	])
+	return [monitor, connector]
+}
 
 export function useDragHandler<
 	DragObject extends DragObjectWithType,
