+++ packages/react-dnd/src/hooks/internal/drop.ts
@@ -7,6 +7,17 @@ import { RefObject, useEffect, useMemo } from 'react'
 import { DropTarget } from 'dnd-core'
 import registerTarget from '../../registerTarget'
 import { useDragDropManager } from './useDragDropManager'
+import createTargetConnector from '../../createTargetConnector'
+import DropTargetMonitorImpl from '../../DropTargetMonitorImpl'
+
+export function useDropTargetMonitor(): [DropTargetMonitor, any] {
+	const manager = useDragDropManager()
+	const monitor = useMemo(() => new DropTargetMonitorImpl(manager), [manager])
+	const connector = useMemo(() => createTargetConnector(manager.getBackend()), [
+		manager,
+	])
+	return [monitor, connector]
+}
 
 export function useDropHandler<
 	DragObject extends DragObjectWithType,
